pi/requirements.txt
pyserial
pyaudio
pygame
paramiko
